requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
psycopg2-binary==2.9.10
ipython==8.28.0
